chore: Rearrange package structure under main_writeview
