+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/component/EmotionGridContainer.java
@@ -1,4 +1,4 @@
-package snu.swpp.moment.ui.main_writeview;
+package snu.swpp.moment.ui.main_writeview.component;
 
 import android.graphics.Typeface;
 import android.util.Log;
