+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/slideview/DailyViewFragment.java
@@ -1,4 +1,4 @@
-package snu.swpp.moment.ui.main_writeview.DaySlide;
+package snu.swpp.moment.ui.main_writeview.slideview;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -26,12 +26,11 @@ import snu.swpp.moment.data.source.StoryRemoteDataSource;
 import snu.swpp.moment.databinding.DailyItemBinding;
 import snu.swpp.moment.exception.NoInternetException;
 import snu.swpp.moment.exception.UnauthorizedAccessException;
-import snu.swpp.moment.ui.main_writeview.DailyViewModel;
-import snu.swpp.moment.ui.main_writeview.DailyViewModelFactory;
-import snu.swpp.moment.ui.main_writeview.GetStoryUseCase;
-import snu.swpp.moment.ui.main_writeview.ListViewAdapter;
-import snu.swpp.moment.ui.main_writeview.ListViewItem;
-import snu.swpp.moment.ui.main_writeview.SaveScoreUseCase;
+import snu.swpp.moment.ui.main_writeview.component.ListFooterContainer;
+import snu.swpp.moment.ui.main_writeview.viewmodel.DailyViewModel;
+import snu.swpp.moment.ui.main_writeview.viewmodel.DailyViewModelFactory;
+import snu.swpp.moment.ui.main_writeview.viewmodel.GetStoryUseCase;
+import snu.swpp.moment.ui.main_writeview.viewmodel.SaveScoreUseCase;
 import snu.swpp.moment.ui.main_writeview.uistate.MomentUiState;
 import snu.swpp.moment.ui.main_writeview.uistate.StoryUiState;
 import snu.swpp.moment.utils.TimeConverter;
