+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/slideview/ListViewItem.java
@@ -1,4 +1,4 @@
-package snu.swpp.moment.ui.main_writeview;
+package snu.swpp.moment.ui.main_writeview.slideview;
 
 import java.util.Date;
 import snu.swpp.moment.data.model.MomentPairModel;
