+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/viewmodel/DailyViewModel.java
@@ -1,4 +1,4 @@
-package snu.swpp.moment.ui.main_writeview;
+package snu.swpp.moment.ui.main_writeview.viewmodel;
 
 import android.util.Log;
 import androidx.lifecycle.MutableLiveData;
