+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/viewmodel/TodayViewModelFactory.java
@@ -1,26 +1,30 @@
-package snu.swpp.moment.ui.main_writeview;
+package snu.swpp.moment.ui.main_writeview.viewmodel;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 import snu.swpp.moment.data.repository.AuthenticationRepository;
 import snu.swpp.moment.data.repository.MomentRepository;
+import snu.swpp.moment.data.repository.StoryRepository;
 
-public class DailyViewModelFactory implements ViewModelProvider.Factory {
+public class TodayViewModelFactory implements ViewModelProvider.Factory {
 
     private final AuthenticationRepository authenticationRepository;
     private final MomentRepository momentRepository;
+    private final StoryRepository storyRepository;
     private final GetStoryUseCase getStoryUseCase;
     private final SaveScoreUseCase saveScoreUseCase;
 
-    public DailyViewModelFactory(
+    public TodayViewModelFactory(
         AuthenticationRepository authenticationRepository,
         MomentRepository momentRepository,
+        StoryRepository storyRepository,
         GetStoryUseCase getStoryUseCase,
         SaveScoreUseCase saveScoreUseCase
     ) {
         this.authenticationRepository = authenticationRepository;
         this.momentRepository = momentRepository;
+        this.storyRepository = storyRepository;
         this.getStoryUseCase = getStoryUseCase;
         this.saveScoreUseCase = saveScoreUseCase;
     }
@@ -28,14 +32,11 @@ public class DailyViewModelFactory implements ViewModelProvider.Factory {
     @NonNull
     @Override
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-        if (modelClass.isAssignableFrom(DailyViewModel.class)) {
-            return (T) new DailyViewModel(
-                authenticationRepository,
-                momentRepository,
-                getStoryUseCase,
-                saveScoreUseCase
-            );
+        if (modelClass.isAssignableFrom(TodayViewModel.class)) {
+            return (T) new TodayViewModel(authenticationRepository, momentRepository,
+                storyRepository, getStoryUseCase, saveScoreUseCase);
         } else {
+            System.out.println("#DEBUG: WriteViewModelFactory not working");
             throw new IllegalArgumentException("Unknown ViewModel class");
         }
     }
