+++ frontend/Moment/app/src/test/java/snu/swpp/moment/ui/main_writeview/TodayViewModelTest.java
@@ -55,6 +55,9 @@ import snu.swpp.moment.ui.main_writeview.uistate.CompletionState;
 import snu.swpp.moment.ui.main_writeview.uistate.CompletionStoreResultState;
 import snu.swpp.moment.ui.main_writeview.uistate.MomentUiState;
 import snu.swpp.moment.ui.main_writeview.uistate.StoryUiState;
+import snu.swpp.moment.ui.main_writeview.viewmodel.GetStoryUseCase;
+import snu.swpp.moment.ui.main_writeview.viewmodel.SaveScoreUseCase;
+import snu.swpp.moment.ui.main_writeview.viewmodel.TodayViewModel;
 
 @RunWith(MockitoJUnitRunner.class)
 public class TodayViewModelTest {
